XFree86: sdksyms: Remove unused -DXorgLoader
We no longer have anything in the tree that checks for XorgLoader.  This
was a fairly monumental hack: xvdi.h used to hide all its functions
behind #ifndef XorgLoader, solely to avoid sdksyms.sh picking up its
symbols, as it was previously a module rather than built-in.

This is no longer the case, so we can remove the define.

Signed-off-by: Daniel Stone <daniel@fooishbar.org>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Keith Packard <keithp@keithp.com>